% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{The function is to take a number of a state in US and a year then filter the data, and
plot a map of all the accidents in that state. The input state number must be in the data! (correct state)
The files need to be in the working directory (Check function fars_read above!).}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number of a state in US as integer}

\item{year}{an input year as integer}
}
\value{
Returns a plot of the accidents based on the \code{state.num} and
\code{year} inputs. Returns an error if the state or year do not exist in the
data set.
}
\description{
The function is to take a number of a state in US and a year then filter the data, and
plot a map of all the accidents in that state. The input state number must be in the data! (correct state)
The files need to be in the working directory (Check function fars_read above!).
}
\examples{
\dontrun{
   fars_map_state(45, 2015)
   fars_read_years(2016) #example of warning (return error!)
   fars_read_years(60, 2016) #example of warning (return error!)
}

}
