% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This function is created in order to check if the file exists and then make it as a data frame
and will throw an error if the file does not exist}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{input the name of the file as charater}
}
\value{
fars_read this function returns an input data with data frame format
}
\description{
You can firstly check if the file exists the \code{stop} function
}
\examples{
\dontrun{
   data <- fars_read(file.csv)
}

}
