% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{The function is to take a list of years and returns a list of data
frames with MONTH and year columns based on data in "accident_X.csv.bz2" files.
The files need to be in the working directory (Check function fars_read above!).}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years of integer or string that when the files were generated}
}
\value{
Returns a list of tibbles (data frames) with the same number of rows
as the data in "accident_<year>.csv.bz2" files and two columns - MONTH and
year. Returns NULL and a warning if the file does not exist.
}
\description{
The function is to take a list of years and returns a list of data
frames with MONTH and year columns based on data in "accident_X.csv.bz2" files.
The files need to be in the working directory (Check function fars_read above!).
}
\examples{
\dontrun{
   fars_read_years(2013:2015)
   fars_read_years(list(2013, 2014))
   fars_read_years(2016) #example of warning (return error!)
}

}
