% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{The function creates a filename for a .csv.bz2 file based on the \code{year}
argument in a form "accident_<year>.csv.bz2". It requires a numerical or
integer input otherwise ends with an error.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the input year as integer or string that when the file was generated}
}
\value{
make_filename this function returns a character vector of file name "accident_X.csv.bz2"
where X will be replaced by the year
}
\description{
The function creates a filename for a .csv.bz2 file based on the \code{year}
argument in a form "accident_<year>.csv.bz2". It requires a numerical or
integer input otherwise ends with an error.
}
\examples{
\dontrun{
   make_filename(2015)
}

}
